Moves folders for incident priority and type UI components inside incident
