+++ src/dispatch/static/dispatch/src/dashboard/incident/IncidentDialogFilter.vue
@@ -58,8 +58,8 @@ import subMonths from "date-fns/subMonths"
 
 import DateWindowInput from "@/components/DateWindowInput.vue"
 import IncidentApi from "@/incident/api"
-import IncidentPriorityCombobox from "@/incident_priority/IncidentPriorityCombobox.vue"
-import IncidentTypeCombobox from "@/incident_type/IncidentTypeCombobox.vue"
+import IncidentPriorityCombobox from "@/incident/priority/IncidentPriorityCombobox.vue"
+import IncidentTypeCombobox from "@/incident/type/IncidentTypeCombobox.vue"
 import ProjectCombobox from "@/project/ProjectCombobox.vue"
 import RouterUtils from "@/router/utils"
 import SearchUtils from "@/search/utils"
