+++ src/dispatch/static/dispatch/src/incident/DetailsTab.vue
@@ -123,8 +123,8 @@ import { required } from "vee-validate/dist/rules"
 import CaseFilterCombobox from "@/case/CaseFilterCombobox.vue"
 import DateTimePickerMenu from "@/components/DateTimePickerMenu.vue"
 import IncidentFilterCombobox from "@/incident/IncidentFilterCombobox.vue"
-import IncidentPrioritySelect from "@/incident_priority/IncidentPrioritySelect.vue"
-import IncidentTypeSelect from "@/incident_type/IncidentTypeSelect.vue"
+import IncidentPrioritySelect from "@/incident/priority/IncidentPrioritySelect.vue"
+import IncidentTypeSelect from "@/incident/type/IncidentTypeSelect.vue"
 import ParticipantSelect from "@/incident/ParticipantSelect.vue"
 import ProjectSelect from "@/project/ProjectSelect.vue"
 import TagFilterAutoComplete from "@/tag/TagFilterAutoComplete.vue"
