+++ src/dispatch/static/dispatch/src/incident/ReportSubmissionForm.vue
@@ -54,11 +54,12 @@
 </template>
 
 <script>
-import { mapFields } from "vuex-map-fields"
 import { ValidationObserver, ValidationProvider, extend } from "vee-validate"
+import { mapFields } from "vuex-map-fields"
 import { required } from "vee-validate/dist/rules"
-import IncidentTypeSelect from "@/incident_type/IncidentTypeSelect.vue"
-import IncidentPrioritySelect from "@/incident_priority/IncidentPrioritySelect.vue"
+
+import IncidentPrioritySelect from "@/incident/priority/IncidentPrioritySelect.vue"
+import IncidentTypeSelect from "@/incident/type/IncidentTypeSelect.vue"
 import ProjectSelect from "@/project/ProjectSelect.vue"
 import TagFilterAutoComplete from "@/tag/TagFilterAutoComplete.vue"
