+++ src/dispatch/static/dispatch/src/incident/Table.vue
@@ -116,8 +116,8 @@ import BulkEditSheet from "@/incident/BulkEditSheet.vue"
 import DeleteDialog from "@/incident/DeleteDialog.vue"
 import IncidentCostCard from "@/incident_cost/IncidentCostCard.vue"
 import IncidentParticipant from "@/incident/Participant.vue"
-import IncidentPriority from "@/incident/IncidentPriority.vue"
-import IncidentStatus from "@/incident/IncidentStatus.vue"
+import IncidentPriority from "@/incident/priority/IncidentPriority.vue"
+import IncidentStatus from "@/incident/status/IncidentStatus.vue"
 import NewSheet from "@/incident/NewSheet.vue"
 import ReportDialog from "@/incident/ReportDialog.vue"
 import RouterUtils from "@/router/utils"
