+++ src/dispatch/static/dispatch/src/incident/TableExportDialog.vue
@@ -106,11 +106,11 @@ import Util from "@/util"
 
 import DateWindowInput from "@/components/DateWindowInput.vue"
 import IncidentApi from "@/incident/api"
-import IncidentPriority from "@/incident/IncidentPriority.vue"
-import IncidentPriorityCombobox from "@/incident_priority/IncidentPriorityCombobox.vue"
-import IncidentStatus from "@/incident/IncidentStatus.vue"
-import IncidentStatusMultiSelect from "@/incident/IncidentStatusMultiSelect.vue"
-import IncidentTypeCombobox from "@/incident_type/IncidentTypeCombobox.vue"
+import IncidentPriority from "@/incident/priority/IncidentPriority.vue"
+import IncidentPriorityCombobox from "@/incident/priority/IncidentPriorityCombobox.vue"
+import IncidentStatus from "@/incident/status/IncidentStatus.vue"
+import IncidentStatusMultiSelect from "@/incident/status/IncidentStatusMultiSelect.vue"
+import IncidentTypeCombobox from "@/incident/type/IncidentTypeCombobox.vue"
 import ProjectCombobox from "@/project/ProjectCombobox.vue"
 import TagFilterAutoComplete from "@/tag/TagFilterAutoComplete.vue"
 import TagTypeFilterCombobox from "@/tag_type/TagTypeFilterCombobox.vue"
