+++ src/dispatch/static/dispatch/src/incident/TableFilterDialog.vue
@@ -64,9 +64,9 @@ import { sum } from "lodash"
 import { mapFields } from "vuex-map-fields"
 
 import DateWindowInput from "@/components/DateWindowInput.vue"
-import IncidentPriorityCombobox from "@/incident_priority/IncidentPriorityCombobox.vue"
-import IncidentStatusMultiSelect from "@/incident/IncidentStatusMultiSelect.vue"
-import IncidentTypeCombobox from "@/incident_type/IncidentTypeCombobox.vue"
+import IncidentPriorityCombobox from "@/incident/priority/IncidentPriorityCombobox.vue"
+import IncidentStatusMultiSelect from "@/incident/status/IncidentStatusMultiSelect.vue"
+import IncidentTypeCombobox from "@/incident/type/IncidentTypeCombobox.vue"
 import ProjectCombobox from "@/project/ProjectCombobox.vue"
 import TagFilterAutoComplete from "@/tag/TagFilterAutoComplete.vue"
 import TagTypeFilterCombobox from "@/tag_type/TagTypeFilterCombobox.vue"
