+++ src/dispatch/static/dispatch/src/incident/priority/store.js
@@ -2,25 +2,22 @@ import { getField, updateField } from "vuex-map-fields"
 import { debounce } from "lodash"
 
 import SearchUtils from "@/search/utils"
-import IncidentTypeApi from "@/incident_type/api"
+import IncidentPriorityApi from "@/incident/priority/api"
 
 const getDefaultSelectedState = () => {
   return {
-    id: null,
-    slug: null,
-    name: null,
-    description: null,
-    visibility: null,
-    incident_template_document: null,
-    tracking_template_document: null,
-    review_template_document: null,
-    executive_template_document: null,
-    loading: false,
-    plugin_metadata: [],
-    exclude_from_metrics: null,
-    enabled: false,
+    color: null,
     default: false,
+    description: null,
+    enabled: false,
+    executive_report_reminder: null,
+    id: null,
+    loading: false,
+    name: null,
+    page_commander: null,
     project: null,
+    tactical_report_reminder: null,
+    view_order: null,
   }
 }
 
@@ -41,7 +38,7 @@ const state = {
       q: "",
       page: 1,
       itemsPerPage: 10,
-      sortBy: ["name"],
+      sortBy: ["view_order"],
       descending: [false],
       filters: {
         project: [],
@@ -60,9 +57,9 @@ const actions = {
     commit("SET_TABLE_LOADING", "primary")
     let params = SearchUtils.createParametersFromTableOptions(
       { ...state.table.options },
-      "IncidentType"
+      "IncidentPriority"
     )
-    return IncidentTypeApi.getAll(params)
+    return IncidentPriorityApi.getAll(params)
       .then((response) => {
         commit("SET_TABLE_LOADING", false)
         commit("SET_TABLE_ROWS", response.data)
@@ -71,15 +68,15 @@ const actions = {
         commit("SET_TABLE_LOADING", false)
       })
   }, 500),
-  createEditShow({ commit }, incidentType) {
+  createEditShow({ commit }, incidentPriority) {
     commit("SET_DIALOG_CREATE_EDIT", true)
-    if (incidentType) {
-      commit("SET_SELECTED", incidentType)
+    if (incidentPriority) {
+      commit("SET_SELECTED", incidentPriority)
     }
   },
-  removeShow({ commit }, incidentType) {
+  removeShow({ commit }, incidentPriority) {
     commit("SET_DIALOG_DELETE", true)
-    commit("SET_SELECTED", incidentType)
+    commit("SET_SELECTED", incidentPriority)
   },
   closeCreateEdit({ commit }) {
     commit("SET_DIALOG_CREATE_EDIT", false)
@@ -91,15 +88,16 @@ const actions = {
   },
   save({ commit, state, dispatch }) {
     commit("SET_SELECTED_LOADING", true)
+
     if (!state.selected.id) {
-      return IncidentTypeApi.create(state.selected)
+      return IncidentPriorityApi.create(state.selected)
         .then(() => {
+          commit("SET_SELECTED_LOADING", false)
           dispatch("closeCreateEdit")
           dispatch("getAll")
-          commit("SET_SELECTED_LOADING", false)
           commit(
             "notification_backend/addBeNotification",
-            { text: "Incident type created successfully.", type: "success" },
+            { text: "Incident priority created successfully.", type: "success" },
             { root: true }
           )
         })
@@ -107,14 +105,14 @@ const actions = {
           commit("SET_SELECTED_LOADING", false)
         })
     } else {
-      return IncidentTypeApi.update(state.selected.id, state.selected)
+      return IncidentPriorityApi.update(state.selected.id, state.selected)
         .then(() => {
           commit("SET_SELECTED_LOADING", false)
           dispatch("closeCreateEdit")
           dispatch("getAll")
           commit(
             "notification_backend/addBeNotification",
-            { text: "Incident type updated successfully.", type: "success" },
+            { text: "Incident priority updated successfully.", type: "success" },
             { root: true }
           )
         })
@@ -124,12 +122,12 @@ const actions = {
     }
   },
   remove({ commit, dispatch }) {
-    return IncidentTypeApi.delete(state.selected.id).then(function () {
+    return IncidentPriorityApi.delete(state.selected.id).then(function () {
       dispatch("closeRemove")
       dispatch("getAll")
       commit(
         "notification_backend/addBeNotification",
-        { text: "Incident type deleted successfully.", type: "success" },
+        { text: "Incident priority deleted successfully.", type: "success" },
         { root: true }
       )
     })
