+++ src/dispatch/static/dispatch/src/incident/type/Table.vue
@@ -4,8 +4,8 @@
     <v-row no-gutters>
       <v-col>
         <v-alert dismissible icon="mdi-school" prominent text type="info"
-          >Priorities adds another dimension to Dispatch's incident categorization. They also allow
-          for some configurability (e.g. only page a command for 'high' priority incidents).
+          >Types categorize incidents. Dispatch allows for configuration on a per-incident type
+          basis.
         </v-alert>
       </v-col>
     </v-row>
@@ -41,9 +41,6 @@
             :loading="loading"
             loading-text="Loading... Please wait"
           >
-            <template v-slot:item.page_commander="{ item }">
-              <v-simple-checkbox v-model="item.page_commander" disabled />
-            </template>
             <template v-slot:item.default="{ item }">
               <v-simple-checkbox v-model="item.default" disabled />
             </template>
@@ -75,11 +72,11 @@
 import { mapFields } from "vuex-map-fields"
 import { mapActions } from "vuex"
 
+import NewEditSheet from "@/incident/type/NewEditSheet.vue"
 import SettingsBreadcrumbs from "@/components/SettingsBreadcrumbs.vue"
-import NewEditSheet from "@/incident_priority/NewEditSheet.vue"
 
 export default {
-  name: "IncidentPriorityTable",
+  name: "IncidentTypeTable",
 
   components: {
     NewEditSheet,
@@ -90,19 +87,17 @@ export default {
       headers: [
         { text: "Name", value: "name", sortable: true },
         { text: "Description", value: "description", sortable: false },
-        { text: "Page Commander", value: "page_commander", sortable: true },
+        { text: "Incident Document", value: "incident_template_document.name", sortable: false },
+        { text: "Visibility", value: "visibility", sortable: false },
         { text: "Default", value: "default", sortable: true },
         { text: "Enabled", value: "enabled", sortable: true },
-        { text: "Tactical Report Reminder", value: "tactical_report_reminder", sortable: true },
-        { text: "Executive Report Reminder", value: "executive_report_reminder", sortable: true },
-        { text: "View Order", value: "view_order", sortable: true },
         { text: "", value: "data-table-actions", sortable: false, align: "end" },
       ],
     }
   },
 
   computed: {
-    ...mapFields("incident_priority", [
+    ...mapFields("incident_type", [
       "table.options.q",
       "table.options.page",
       "table.options.itemsPerPage",
@@ -113,13 +108,20 @@ export default {
       "table.rows.items",
       "table.rows.total",
     ]),
-    ...mapFields("route", ["query", "params"]),
+    ...mapFields("route", ["query"]),
   },
 
   created() {
     this.project = [{ name: this.query.project }]
 
     this.getAll()
+
+    this.$watch(
+      (vm) => [vm.page],
+      () => {
+        this.getAll()
+      }
+    )
 
     this.$watch(
       (vm) => [vm.q, vm.itemsPerPage, vm.sortBy, vm.descending, vm.project],
@@ -132,7 +134,7 @@ export default {
   },
 
   methods: {
-    ...mapActions("incident_priority", ["getAll", "createEditShow", "removeShow"]),
+    ...mapActions("incident_type", ["getAll", "createEditShow", "removeShow"]),
   },
 }
 </script>
