+++ src/dispatch/static/dispatch/src/search/SearchFilterCreateDialog.vue
@@ -167,21 +167,20 @@
 </template>
 
 <script>
-import { mapFields } from "vuex-map-fields"
 import { ValidationObserver, ValidationProvider, extend } from "vee-validate"
+import { mapActions } from "vuex"
+import { mapFields } from "vuex-map-fields"
 import { required } from "vee-validate/dist/rules"
 
-import { mapActions } from "vuex"
-
 import IncidentApi from "@/incident/api"
-import IncidentStatusMultiSelect from "@/incident/IncidentStatusMultiSelect.vue"
-import TagFilterAutoComplete from "@/tag/TagFilterAutoComplete.vue"
-import IncidentTypeCombobox from "@/incident_type/IncidentTypeCombobox.vue"
-import IncidentPriorityCombobox from "@/incident_priority/IncidentPriorityCombobox.vue"
-import TagTypeFilterCombobox from "@/tag_type/TagTypeFilterCombobox.vue"
-import IncidentStatus from "@/incident/IncidentStatus.vue"
-import IncidentPriority from "@/incident/IncidentPriority.vue"
+import IncidentPriority from "@/incident/priority/IncidentPriority.vue"
+import IncidentPriorityCombobox from "@/incident/priority/IncidentPriorityCombobox.vue"
+import IncidentStatus from "@/incident/status/IncidentStatus.vue"
+import IncidentStatusMultiSelect from "@/incident/status/IncidentStatusMultiSelect.vue"
+import IncidentTypeCombobox from "@/incident/type/IncidentTypeCombobox.vue"
 import SearchUtils from "@/search/utils"
+import TagFilterAutoComplete from "@/tag/TagFilterAutoComplete.vue"
+import TagTypeFilterCombobox from "@/tag_type/TagTypeFilterCombobox.vue"
 
 extend("required", {
   ...required,
